Add swith to OrgUnitsSelector to use shortnames.
Reference ticket: https://app.clickup.com/t/2ghev8r

We want the OU tree to be able to show shortnames instead of just the
default displaynames. This commit adds a Switch inside a FormControlLabel
between the search box and the OU tree of the OrgUnitsSelector
component to do just that.

The trickiest part was adding:
  key={useShortNames}
to the OrgUnitTree component, because it not having a unique key before
meant that react never knew it had to update it after changing the
state of the OrgUnitsSelector component.